pyproject.toml
[project]
name = "pega-vagas"
version = "0.1.0"
description = "Pipeline de Engenharia de Dados para scraping de vagas de tecnologia"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }

dependencies = [
    # Web Scraping
    "camoufox>=0.4.0",
    "playwright>=1.40.0",
    
    # LLM
    "google-generativeai>=0.8.0",
    "openai>=1.50.0",
    
    # Data Processing
    "duckdb>=1.1.0",
    "pandas>=2.2.0",
    "pyarrow>=18.0.0",
    
    # Validation
    "pydantic>=2.9.0",
    
    # Utils
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "tenacity>=9.0.0",
    "trafilatura>=1.12.0",
    "structlog>=24.0.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["src"]
